move AddEmployee to component's features folder
