+++ src/components/features/employees/AddEmployee.tsx
@@ -2,20 +2,21 @@ import {
   Button,
   Card,
   CardContent,
+  CircularProgress,
   FormControl,
   Grid2,
   TextField,
   Typography,
 } from "@mui/material";
 import React from "react";
-import theme from "../../theme";
-import API_KEYS from "../../api/keys";
-import useApiRequest from "../../hooks/useApiRequest";
-import { Navigate, useNavigate } from "react-router";
+import theme from "../../../theme";
+import API_KEYS from "../../../api/keys";
+import useApiRequest from "../../../hooks/useApiRequest";
+import { useNavigate } from "react-router";
 
 const AddEmployee = () => {
-    const navigate = useNavigate()
-  const { data, loading, error, getData } = useApiRequest(
+  const navigate = useNavigate();
+  const { loading, error, getData } = useApiRequest(
     `${import.meta.env.VITE_HOST_URL}${API_KEYS.employees}`,
     "POST"
   );
@@ -32,10 +33,18 @@ const AddEmployee = () => {
         salary: e.target.elements.salary.value,
       }),
     });
-    navigate("/employees")
+    navigate("/employees");
   };
+
+  if(loading){
+    return <CircularProgress />
+  }
+
+  if(error){
+    return <Typography>{error.message}</Typography>
+  }
+
   return (
-    
     <Grid2 size={{ xs: 12, sm: 6, md: 4 }}>
       <Card variant="outlined">
         <CardContent>
@@ -51,46 +60,22 @@ const AddEmployee = () => {
           </Typography>
           <form onSubmit={handleFormSubmit}>
             <FormControl sx={{ mb: "16px" }} fullWidth>
-              <TextField
-                id="firstName"
-                placeholder="First Name "
-                type="text"
-              />
+              <TextField id="firstName" placeholder="First Name " type="text" />
             </FormControl>
             <FormControl sx={{ mb: "16px" }} fullWidth>
-              <TextField
-                id="lastName"
-                placeholder="LastName"
-                type="text"
-              />
+              <TextField id="lastName" placeholder="LastName" type="text" />
             </FormControl>
             <FormControl sx={{ mb: "16px" }} fullWidth>
-              <TextField
-                id="email"
-                placeholder="Email"
-                type="email"
-              />
+              <TextField id="email" placeholder="Email" type="email" />
             </FormControl>
             <FormControl sx={{ mb: "16px" }} fullWidth>
-              <TextField
-                id="position"
-                placeholder="Position"
-                type="text"
-              />
+              <TextField id="position" placeholder="Position" type="text" />
             </FormControl>
             <FormControl sx={{ mb: "16px" }} fullWidth>
-              <TextField
-                id="department"
-                placeholder="Department"
-                type="text"
-              />
+              <TextField id="department" placeholder="Department" type="text" />
             </FormControl>
             <FormControl sx={{ mb: "16px" }} fullWidth>
-              <TextField
-                id="salary"
-                placeholder="Salary"
-                type="number"
-              />
+              <TextField id="salary" placeholder="Salary" type="number" />
             </FormControl>
             <FormControl sx={{ mb: "16px" }} fullWidth>
               <Button type="submit" variant="contained">
@@ -98,6 +83,7 @@ const AddEmployee = () => {
               </Button>
             </FormControl>
           </form>
+
         </CardContent>
       </Card>
     </Grid2>
